sfesef

+++ first.c
@@ -2,37 +2,37 @@
 
 int func1 (void)
 {
-    double eps = 1;
+    __float128 eps = 1;
 #pragma GCC diagnostic push
 #pragma GCC diagnostic ignored "-Wfloat-equal"
     for (int i = 0;; i++)
     {
-        if (1. + eps == 1.)        //-Wfloat-equal
+        if ((__float128)1. + eps == (__float128)1.)        //-Wfloat-equal
         {
 #pragma GCC diagnostic pop
             return i;
         }
-        eps = ldexp (1., -i);
+        eps = (__float128)ldexpl ((long double)1., -i);
     }
     return 0;
 }
 
-double func2 (void)
+__float128 func2 (void)
 {
-    double x = 0.;
+    __float128 x = (__float128)0.;
 #pragma GCC diagnostic push
 #pragma GCC diagnostic ignored "-Wfloat-equal"
-    for (int i = 0; x + 1. != x; i++) { x += ldexp (2., i / 20); }
+    for (int i = 0; x + (__float128)1. != x; i++) { x += (__float128)ldexpl ((long double)2., i / 20); }
 #pragma GCC diagnostic pop
     return x;
 }
 
-double func3 (void)
+__float128 func3 (void)
 {
-    double x = 0.;
+    __float128 x = (__float128)0.;
 #pragma GCC diagnostic push
 #pragma GCC diagnostic ignored "-Wfloat-equal"
-    for (int i = 0; x + pow (10., 20) != x; i++) { x += ldexp (2., i / 20); }
+    for (int i = 0; x + (__float128)pow (10., 20) != x; i++) { x += (__float128)ldexpl ((long double)2., i / 20); }
 #pragma GCC diagnostic pop
     return x;
 }
